TECH Remove the Fields alias
I started to use this new logger lib in beqos, and then I realized that as soon
as I started to use the child logger feature of logrus, my code began to mix
calls to logger.WithFields(logger.Fields) and logrus.Logger.WithFields(logrus.Fields)
in a very awkward way.

My first reaction was to think about extending the Logger type with our own, so
I could have a logger.Logger.WithFields(logger.Fields) methods, but then I remembered
this excellent article, and especially section 4 about shallow wrappers
https://medium.com/@rdsubhas/10-modern-software-engineering-mistakes-bc67fbef4fc8#.ykinlwuce

So we need to remember the original purpose of this lib, which is *NOT* to
wrap logrus to switch in an hypothetical future, but to add a thin configuration layer
for logentries and sentry, and possibly any other future log destination. So the only
reasonable thing to do is to remove the Fields alias.